- server moved to a new dir server - and the separate package json is made now
