+++ package.json
@@ -13,8 +13,6 @@
     "@hookform/resolvers": "^4.1.3",
     "@tailwindcss/vite": "^4.1.3",
     "axios": "^1.8.4",
-    "bun": "^1.2.8",
-    "nodemailer": "^6.10.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-hook-form": "^7.55.0",
@@ -25,12 +23,10 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.24.0",
-    "@types/bun": "^1.2.8",
-    "@types/nodemailer": "^6.4.17",
     "@types/react": "^19.1.0",
-    "@types/react-dom": "^19.1.1",
+    "@types/react-dom": "^19.1.2",
     "@vitejs/plugin-react-swc": "^3.8.1",
-    "daisyui": "^5.0.16",
+    "daisyui": "^5.0.18",
     "eslint": "^9.24.0",
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.19",
@@ -39,6 +35,6 @@
     "prettier-plugin-tailwindcss": "^0.6.11",
     "typescript": "~5.7.3",
     "typescript-eslint": "^8.29.1",
-    "vite": "^6.2.5"
+    "vite": "^6.2.6"
   }
 }
