+++ server/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "server",
+  "module": "index.ts",
+  "type": "module",
+  "private": true,
+  "devDependencies": {
+    "@types/bun": "^1.2.9"
+  },
+  "peerDependencies": {
+    "typescript": "^5.8.3"
+  },
+  "dependencies": {
+    "@types/nodemailer": "^6.4.17",
+    "bun": "^1.2.9",
+    "nodemailer": "^6.10.0"
+  }
+}
